Players can be added and players added to gameplay

+++ gameplay.py
@@ -3,6 +3,7 @@ import cv2 as cv
 import os
 from difference import *
 from tip import *
+import time
 
 def startGame(boardPoints):
 
@@ -25,8 +26,9 @@ def startGame(boardPoints):
         frameA = stream1.read()
         frameB = stream2.read()
 
-        cv.imwrite('Images/emptyA.jpg', frameA)
-        cv.imwrite('Images/emptyB.jpg', frameB)
+        if not os.path.isfile('Images/emptyA.jpg'):
+            cv.imwrite('Images/emptyA.jpg', frameA)
+            cv.imwrite('Images/emptyB.jpg', frameB)
 
         if frameA is None or frameB is None:
             break
@@ -80,6 +82,38 @@ def startGame(boardPoints):
         
 
     cv.destroyAllWindows()
+
+    stream1.stop()
+    stream2.stop()
+
+def waitForEmptyBoard():
+    stream1 = VideoGear(source=0, logging=True).start() 
+    stream2 = VideoGear(source=1, logging=True).start() 
+
+    emptyBoardA = cv.imread("Images/emptyA.jpg")
+    emptyBoardB = cv.imread("Images/emptyB.jpg")
+
+    for i in range(1, 10):
+        frameA = stream1.read()
+        frameB = stream2.read()
+    
+    cv.imwrite("Images/frameA.jpg", frameA)
+    cv.imwrite("Images/frameB.jpg", frameB)
+
+    cntsA, boardContoursA, contourFoundA = findContour(emptyBoardA, frameA, 10, "A")
+    cntsB, boardContoursB, contourFoundB = findContour(emptyBoardB, frameB, 10, "B")
+
+    while contourFoundA or contourFoundB:
+        time.sleep(0.2)
+
+        frameA = stream1.read()
+        frameB = stream2.read()
+
+        cntsA, boardContoursA, contourFoundA = findContour(emptyBoardA, frameA, 200, "A")
+        cntsB, boardContoursB, contourFoundB = findContour(emptyBoardB, frameB, 200, "B")
+    
+    os.remove("Images/emptyA.jpg")
+    os.remove("Images/emptyB.jpg")
 
     stream1.stop()
     stream2.stop()
